Make code coverage ignore *-ui.R files
UI elements are really difficult to automatically test. The 'best'
way to do so is by making a screenshot and comparing it to a
previous screenshot. A single pixel being different causes the
test to fail. Secondly, UI is much more prone to change than the
backend is. This means these screenshot tests often fail while
everything is actually working just fine.

Because testing is hard for UI, I do not expect them to be tested.
It therefore does not make sense to check coverage for UI either.
This is why I opt to ignore all *-ui.R files for code coverage.