Add shinydashboard for nicer UI
This patch adds shinydashboard as a package, and alters the
main_ui.R to make use of this package. It makes the dashboard look
like the dashboard linked in #25. It also adds a line in run.R,
which lets us make static files like .css or .js files to the web
server shiny starts. Just put them in the inst/www/ folder and the
files will be added under <shiny_server_url>/www/.