Fix warnings on previous R version
In previous versions of R, data.frame defaults to changing strings
into factors. As of R 4.0, this is no longer the case. This is why
builds on Travis are currently failing, because it tests both on
R 4.0 and R 3.X. This patch explicitly sets the stringsToFactors to
false in the offending data.frame call.